requirements.txt
cloudinary==1.36.0
dj-database-url==2.1.0
Django==4.2.6
django-blog-zinnia==0.20
django-ckeditor==6.7.0
django-classy-tags==4.1.0
django-cloudinary-storage==0.3.0
gunicorn==21.2.0
Mezzanine==6.0.0
psycopg2==2.9.9
requests==2.31.0
whitenoise==6.6.0
mysqlclient